handle multi_run_config in sim_config yaml and code clean up in actuated_cylinder_diffusion
