+++ planning_through_contact/simulation/planar_pushing/planar_pushing_sim_config.py
@@ -28,14 +28,66 @@ from planning_through_contact.geometry.planar.planar_pushing_trajectory import (
 )
 from planning_through_contact.geometry.rigid_body import RigidBody
 from planning_through_contact.planning.planar.planar_plan_config import (
-    SliderPusherSystemConfig, 
-    MultiRunConfig
+    SliderPusherSystemConfig,
+    BoxWorkspace,
+    PlanarPushingWorkspace,
 )
 from planning_through_contact.simulation.controllers.hybrid_mpc import HybridMpcConfig
 from planning_through_contact.experiments.utils import (
     get_box,
     get_tee,
+    get_default_plan_config,
+    get_slider_start_poses,
 )
+
+class MultiRunConfig:
+    def __init__(
+            self,
+            num_runs, 
+            max_attempt_duration, 
+            seed, 
+            slider_type,
+            pusher_start_pose,
+            slider_goal_pose,
+            workspace_width=0.35,
+            workspace_height=0.5,
+    ):
+        # Set up multi run config
+        config = get_default_plan_config(
+            slider_type=slider_type,
+            pusher_radius=0.015,
+            hardware=False,
+        )
+        # update config (probably don't need these)
+        config.contact_config.lam_min = 0.15
+        config.contact_config.lam_max = 0.85
+        config.non_collision_cost.distance_to_object_socp = 0.25   
+
+        # Get initial slider poses
+        workspace = PlanarPushingWorkspace(
+            slider=BoxWorkspace(
+                width=workspace_width,
+                height=workspace_height,
+                center=np.array([slider_goal_pose.x, slider_goal_pose.y]),
+                buffer=0,
+            ),
+        )
+        self.initial_slider_poses = get_slider_start_poses(
+            seed=seed,
+            num_plans=num_runs,
+            workspace=workspace,
+            config=config,
+            pusher_pose=pusher_start_pose,
+            limit_rotations=False,
+        )
+        self.target_slider_poses = [slider_goal_pose] * num_runs
+        self.max_attempt_duration = max_attempt_duration
+    
+    def __str__(self):
+        slider_pose_str = f"initial_slider_poses: {self.initial_slider_poses}"
+        target_pose_str = f"target_slider_poses: {self.target_slider_poses}"
+        return f"{slider_pose_str}\n{target_pose_str}\nmax_attempt_duration: {self.max_attempt_duration}"
+
 
 @dataclass
 class PlanarPushingSimConfig:
@@ -84,20 +136,42 @@ class PlanarPushingSimConfig:
 
     @classmethod
     def from_yaml(cls, cfg: OmegaConf):
+        # Create sim_config with mandatory fields
         # TODO: read slider directly from yaml instead of if statement
         slider: RigidBody = get_box() if cfg.slider_type == "box" else get_tee()
         dynamics_config: SliderPusherSystemConfig = hydra.utils.instantiate(
             cfg.dynamics_config,
         )
         dynamics_config.slider = slider
-
         slider_goal_pose: PlanarPose = hydra.utils.instantiate(cfg.slider_goal_pose)
-        pusher_start_pose: PlanarPose = hydra.utils.instantiate(cfg.pusher_start_pose)
-        slider_start_pose: PlanarPose = hydra.utils.instantiate(cfg.slider_start_pose)
-        default_joint_positions = np.array(cfg.default_joint_positions)
-        mpc_config: HybridMpcConfig = hydra.utils.instantiate(cfg.mpc_config)
-        
-        camera_configs = None
+        pusher_start_pose: PlanarPose = hydra.utils.instantiate(cfg.pusher_start_pose)   
+        sim_config = cls(
+            dynamics_config=dynamics_config,
+            slider=slider,
+            contact_model=eval(cfg.contact_model),
+            visualize_desired=cfg.visualize_desired,
+            slider_goal_pose=slider_goal_pose,
+            pusher_start_pose=pusher_start_pose,
+            time_step=cfg.time_step,
+            closed_loop=cfg.closed_loop,
+            draw_frames=cfg.draw_frames,
+            use_realtime=cfg.use_realtime,
+            delay_before_execution=cfg.delay_before_execution,
+            save_plots=cfg.save_plots,
+            scene_directive_name=cfg.scene_directive_name,
+            use_hardware=cfg.use_hardware,
+            pusher_z_offset=cfg.pusher_z_offset,
+            collect_data=cfg.collect_data,
+            data_dir=cfg.data_dir,
+        )
+
+        # Optional fields
+        if 'slider_start_pose' in cfg:
+            sim_config.slider_start_pose = hydra.utils.instantiate(cfg.slider_start_pose)
+        if 'default_joint_positions' in cfg:
+            sim_config.default_joint_positions = np.array(cfg.default_joint_positions)
+        if 'mpc_config' in cfg:
+            sim_config.mpc_config = hydra.utils.instantiate(cfg.mpc_config)
         if 'camera_configs' in cfg and cfg.camera_configs:
             camera_configs = []
             for camera_config in cfg.camera_configs:
@@ -121,33 +195,8 @@ class PlanarPushingSimConfig:
                         show_rgb=camera_config.show_rgb,
                     )
                 )
-        
+            sim_config.camera_configs = camera_configs
         if 'multi_run_config' in cfg and cfg.multi_run_config:
-            multi_run_config = hydra.utils.instantiate(cfg.multi_run_config)
-        else:
-            multi_run_config = None
+            sim_config.multi_run_config = hydra.utils.instantiate(cfg.multi_run_config)
 
-        return cls(
-            dynamics_config=dynamics_config,
-            slider=slider,
-            contact_model=eval(cfg.contact_model),
-            visualize_desired=cfg.visualize_desired,
-            slider_goal_pose=slider_goal_pose,
-            pusher_start_pose=pusher_start_pose,
-            slider_start_pose=slider_start_pose,
-            default_joint_positions=default_joint_positions,
-            time_step=cfg.time_step,
-            closed_loop=cfg.closed_loop,
-            draw_frames=cfg.draw_frames,
-            use_realtime=cfg.use_realtime,
-            delay_before_execution=cfg.delay_before_execution,
-            save_plots=cfg.save_plots,
-            mpc_config=mpc_config,
-            scene_directive_name=cfg.scene_directive_name,
-            use_hardware=cfg.use_hardware,
-            pusher_z_offset=cfg.pusher_z_offset,
-            camera_configs=camera_configs,
-            collect_data=cfg.collect_data,
-            data_dir=cfg.data_dir,
-            multi_run_config=multi_run_config
-        )
+        return sim_config
